chore: use arena allocator for intermediate values in config.zig
Using std.testing.allocator consistently in all unit tests for config
parsing revealed that src/config.zig previously leaked memory for a
couple of intermediate values. For example default values for config
settings that go out of scope when they are replaced with the actual
configured value from the config file or from environment variables, but
also allocated strings that are handled during parsing.

While it is certainly possible to get rid of those by carefully
sprinkling allocator.free calls all over the code, this approach is
error-prone and will also incur high maintenance effort down the road.

Using an arena allocator for the intermediate values and freeing them
all at once when the parsing is done is more robust.A

This fixes #181 partially, for src/config.zig.